End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { PaginationParams } from '../_models/PaginationParams';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pagParams: PaginationParams = new PaginationParams();
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -34,11 +35,18 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm("Confirm delete message", "This cannot be undone").subscribe({
+      next: result => {
+        if(result){
+          this.messageService.deleteMessage(id).subscribe({
+            next: () => {
+              this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+            }
+          })
+        }
       }
     })
+    
   }
 
   pageChanged(event: any) {
